feat(BoardService): extract board setup and mode-specific game starts
refactor(GameService): delegate board lifecycle to BoardService

test: integrate BoardService into TestServiceFactory
